+++ recommenders/models/sasrec/model.py
@@ -1,8 +1,13 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
+import copy
+import random
 import tensorflow as tf
 import numpy as np
-import sys
+from tqdm import tqdm
+import time
+
+from recommenders.utils.timer import Timer
 
 
 class MultiHeadAttention(tf.keras.layers.Layer):
@@ -34,10 +39,10 @@ class MultiHeadAttention(tf.keras.layers.Layer):
         V = self.V(keys)  # (N, T_k, C)
 
         # --- MULTI HEAD ---
-        # Split and concat
-        Q_ = tf.concat(tf.split(Q, self.num_heads, axis=2), axis=0)  # (h*N, T_q, C/h)
-        K_ = tf.concat(tf.split(K, self.num_heads, axis=2), axis=0)  # (h*N, T_k, C/h)
-        V_ = tf.concat(tf.split(V, self.num_heads, axis=2), axis=0)  # (h*N, T_k, C/h)
+        # Split and concat, Q_, K_ and V_ are all (h*N, T_q, C/h)
+        Q_ = tf.concat(tf.split(Q, self.num_heads, axis=2), axis=0)
+        K_ = tf.concat(tf.split(K, self.num_heads, axis=2), axis=0)
+        V_ = tf.concat(tf.split(V, self.num_heads, axis=2), axis=0)
 
         # --- SCALED DOT PRODUCT ---
         # Multiplication
@@ -54,7 +59,8 @@ class MultiHeadAttention(tf.keras.layers.Layer):
         )  # (h*N, T_q, T_k)
 
         paddings = tf.ones_like(outputs) * (-(2 ** 32) + 1)
-        outputs = tf.where(tf.equal(key_masks, 0), paddings, outputs)  # (h*N, T_q, T_k)
+        # outputs, (h*N, T_q, T_k)
+        outputs = tf.where(tf.equal(key_masks, 0), paddings, outputs)
 
         # Future blinding (Causality)
         diag_vals = tf.ones_like(outputs[0, :, :])  # (T_q, T_k)
@@ -66,13 +72,14 @@ class MultiHeadAttention(tf.keras.layers.Layer):
         )  # (h*N, T_q, T_k)
 
         paddings = tf.ones_like(masks) * (-(2 ** 32) + 1)
-        outputs = tf.where(tf.equal(masks, 0), paddings, outputs)  # (h*N, T_q, T_k)
+        # outputs, (h*N, T_q, T_k)
+        outputs = tf.where(tf.equal(masks, 0), paddings, outputs)
 
         # Activation
         outputs = tf.nn.softmax(outputs)  # (h*N, T_q, T_k)
 
-        # Query Masking
-        query_masks = tf.sign(tf.abs(tf.reduce_sum(queries, axis=-1)))  # (N, T_q)
+        # Query Masking, query_masks (N, T_q)
+        query_masks = tf.sign(tf.abs(tf.reduce_sum(queries, axis=-1)))
         query_masks = tf.tile(query_masks, [self.num_heads, 1])  # (h*N, T_q)
         query_masks = tf.tile(
             tf.expand_dims(query_masks, -1), [1, 1, tf.shape(keys)[1]]
@@ -273,7 +280,8 @@ class SASREC(tf.keras.Model):
         Recommendation. Proceedings of IEEE International Conference on
         Data Mining (ICDM'18)
 
-        Original source code from nnkkmto/SASRec-tf2, https://github.com/nnkkmto/SASRec-tf2
+        Original source code from nnkkmto/SASRec-tf2,
+        https://github.com/nnkkmto/SASRec-tf2
 
     Args:
         item_num: number of items in the dataset
@@ -469,3 +477,241 @@ class SASREC(tf.keras.Model):
         loss += reg_loss
 
         return loss
+
+    def create_combined_dataset(self, u, seq, pos, neg):
+        """
+        function to create model inputs from sampled batch data.
+        This function is used only during training.
+        """
+        inputs = {}
+        seq = tf.keras.preprocessing.sequence.pad_sequences(
+            seq, padding="pre", truncating="pre", maxlen=self.seq_max_len
+        )
+        pos = tf.keras.preprocessing.sequence.pad_sequences(
+            pos, padding="pre", truncating="pre", maxlen=self.seq_max_len
+        )
+        neg = tf.keras.preprocessing.sequence.pad_sequences(
+            neg, padding="pre", truncating="pre", maxlen=self.seq_max_len
+        )
+
+        inputs["users"] = np.expand_dims(np.array(u), axis=-1)
+        inputs["input_seq"] = seq
+        inputs["positive"] = pos
+        inputs["negative"] = neg
+
+        target = np.concatenate(
+            [
+                np.repeat(1, seq.shape[0] * seq.shape[1]),
+                np.repeat(0, seq.shape[0] * seq.shape[1]),
+            ],
+            axis=0,
+        )
+        target = np.expand_dims(target, axis=-1)
+        return inputs, target
+
+    def train(self, dataset, sampler, **kwargs):
+        """
+        High level function for model training as well as
+        evaluation on the validation and test dataset
+        """
+        num_epochs = kwargs.get("num_epochs", 10)
+        batch_size = kwargs.get("batch_size", 128)
+        lr = kwargs.get("learning_rate", 0.001)
+        val_epoch = kwargs.get("val_epoch", 5)
+
+        num_steps = int(len(dataset.user_train) / batch_size)
+
+        optimizer = tf.keras.optimizers.Adam(
+            learning_rate=lr, beta_1=0.9, beta_2=0.999, epsilon=1e-7
+        )
+
+        loss_function = self.loss_function
+
+        train_loss = tf.keras.metrics.Mean(name="train_loss")
+
+        train_step_signature = [
+            {
+                "users": tf.TensorSpec(shape=(None, 1), dtype=tf.int64),
+                "input_seq": tf.TensorSpec(
+                    shape=(None, self.seq_max_len), dtype=tf.int64
+                ),
+                "positive": tf.TensorSpec(
+                    shape=(None, self.seq_max_len), dtype=tf.int64
+                ),
+                "negative": tf.TensorSpec(
+                    shape=(None, self.seq_max_len), dtype=tf.int64
+                ),
+            },
+            tf.TensorSpec(shape=(None, 1), dtype=tf.int64),
+        ]
+
+        @tf.function(input_signature=train_step_signature)
+        def train_step(inp, tar):
+            with tf.GradientTape() as tape:
+                pos_logits, neg_logits, loss_mask = self(inp, training=True)
+                loss = loss_function(pos_logits, neg_logits, loss_mask)
+
+            gradients = tape.gradient(loss, self.trainable_variables)
+            optimizer.apply_gradients(zip(gradients, self.trainable_variables))
+
+            train_loss(loss)
+            return loss
+
+        T = 0.0
+        t0 = Timer()
+        t0.start()
+
+        for epoch in range(1, num_epochs + 1):
+
+            step_loss = []
+            train_loss.reset_states()
+            for step in tqdm(
+                range(num_steps), total=num_steps, ncols=70, leave=False, unit="b"
+            ):
+
+                u, seq, pos, neg = sampler.next_batch()
+
+                inputs, target = self.create_combined_dataset(u, seq, pos, neg)
+
+                loss = train_step(inputs, target)
+                step_loss.append(loss)
+
+            if epoch % val_epoch == 0:
+                t0.stop()
+                t1 = t0.interval
+                T += t1
+                print("Evaluating...")
+                t_test = self.evaluate(dataset)
+                t_valid = self.evaluate_valid(dataset)
+                print(
+                    f"\nepoch: {epoch}, time: {T}, valid (NDCG@10: {t_valid[0]}, HR@10: {t_valid[1]})"
+                )
+                print(
+                    f"epoch: {epoch}, time: {T},  test (NDCG@10: {t_test[0]}, HR@10: {t_test[1]})"
+                )
+                t0.start()
+
+        t_test = self.evaluate(dataset)
+        print(f"\nepoch: {epoch}, test (NDCG@10: {t_test[0]}, HR@10: {t_test[1]})")
+
+        return t_test
+
+    def evaluate(self, dataset):
+        """
+        Evaluation on the test users (users with at least 3 items)
+        """
+        usernum = dataset.usernum
+        itemnum = dataset.itemnum
+        train = dataset.user_train  # removing deepcopy
+        valid = dataset.user_valid
+        test = dataset.user_test
+
+        NDCG = 0.0
+        HT = 0.0
+        valid_user = 0.0
+
+        if usernum > 10000:
+            users = random.sample(range(1, usernum + 1), 10000)
+        else:
+            users = range(1, usernum + 1)
+
+        for u in tqdm(users, ncols=70, leave=False, unit="b"):
+
+            if len(train[u]) < 1 or len(test[u]) < 1:
+                continue
+
+            seq = np.zeros([self.seq_max_len], dtype=np.int32)
+            idx = self.seq_max_len - 1
+            seq[idx] = valid[u][0]
+            idx -= 1
+            for i in reversed(train[u]):
+                seq[idx] = i
+                idx -= 1
+                if idx == -1:
+                    break
+            rated = set(train[u])
+            rated.add(0)
+            item_idx = [test[u][0]]
+            for _ in range(self.num_neg_test):
+                t = np.random.randint(1, itemnum + 1)
+                while t in rated:
+                    t = np.random.randint(1, itemnum + 1)
+                item_idx.append(t)
+
+            inputs = {}
+            inputs["user"] = np.expand_dims(np.array([u]), axis=-1)
+            inputs["input_seq"] = np.array([seq])
+            inputs["candidate"] = np.array([item_idx])
+
+            # inverse to get descending sort
+            predictions = -1.0 * self.predict(inputs)
+            predictions = np.array(predictions)
+            predictions = predictions[0]
+
+            rank = predictions.argsort().argsort()[0]
+
+            valid_user += 1
+
+            if rank < 10:
+                NDCG += 1 / np.log2(rank + 2)
+                HT += 1
+
+        return NDCG / valid_user, HT / valid_user
+
+    def evaluate_valid(self, dataset):
+        """
+        Evaluation on the validation users
+        """
+        usernum = dataset.usernum
+        itemnum = dataset.itemnum
+        train = dataset.user_train  # removing deepcopy
+        valid = dataset.user_valid
+
+        NDCG = 0.0
+        valid_user = 0.0
+        HT = 0.0
+        if usernum > 10000:
+            users = random.sample(range(1, usernum + 1), 10000)
+        else:
+            users = range(1, usernum + 1)
+
+        for u in tqdm(users, ncols=70, leave=False, unit="b"):
+            if len(train[u]) < 1 or len(valid[u]) < 1:
+                continue
+
+            seq = np.zeros([self.seq_max_len], dtype=np.int32)
+            idx = self.seq_max_len - 1
+            for i in reversed(train[u]):
+                seq[idx] = i
+                idx -= 1
+                if idx == -1:
+                    break
+
+            rated = set(train[u])
+            rated.add(0)
+            item_idx = [valid[u][0]]
+            for _ in range(self.num_neg_test):
+                t = np.random.randint(1, itemnum + 1)
+                while t in rated:
+                    t = np.random.randint(1, itemnum + 1)
+                item_idx.append(t)
+
+            inputs = {}
+            inputs["user"] = np.expand_dims(np.array([u]), axis=-1)
+            inputs["input_seq"] = np.array([seq])
+            inputs["candidate"] = np.array([item_idx])
+
+            # predictions = -model.predict(sess, [u], [seq], item_idx)
+            predictions = -1.0 * self.predict(inputs)
+            predictions = np.array(predictions)
+            predictions = predictions[0]
+
+            rank = predictions.argsort().argsort()[0]
+
+            valid_user += 1
+
+            if rank < 10:
+                NDCG += 1 / np.log2(rank + 2)
+                HT += 1
+
+        return NDCG / valid_user, HT / valid_user
